Replaced RingClient interface with concrete class
