+++ src/main/java/com/appsflyer/donkey/client/ring/RingClient.java
@@ -18,21 +18,82 @@
 package com.appsflyer.donkey.client.ring;
 
 import clojure.lang.IPersistentMap;
-import com.appsflyer.donkey.client.Client;
 import com.appsflyer.donkey.client.ClientConfig;
 import io.vertx.core.Future;
+import io.vertx.core.MultiMap;
+import io.vertx.core.Promise;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.ext.web.client.HttpRequest;
+import io.vertx.ext.web.client.WebClient;
+import io.vertx.ext.web.multipart.MultipartForm;
 
-public interface RingClient extends Client<IPersistentMap> {
+import static com.appsflyer.donkey.util.TypeConverter.*;
+
+@SuppressWarnings("WeakerAccess")
+public final class RingClient {
   
-  static RingClient create(ClientConfig config) {
-    return new RingClientImpl(config);
+  public static RingClient create(ClientConfig config) {
+    return new RingClient(config);
   }
   
-  Future<IPersistentMap> send(HttpRequest<Buffer> request, Object body);
+  private final WebClient client;
+  private final RequestFactory requestFactory;
   
-  Future<IPersistentMap> sendForm(HttpRequest<Buffer> request, IPersistentMap body);
+  private RingClient(ClientConfig config) {
+    client = WebClient.create(config.vertx(), config.clientOptions());
+    requestFactory = new RequestFactory(client);
+  }
   
-  Future<IPersistentMap> sendMultiPartForm(HttpRequest<Buffer> request, IPersistentMap body);
+  public HttpRequest<Buffer> request(IPersistentMap opts) {
+    return requestFactory.create(opts);
+  }
+  
+  public Future<IPersistentMap> send(HttpRequest<Buffer> request) {
+    Promise<IPersistentMap> promise = Promise.promise();
+    request.send(new RingResponseAdapter(promise));
+    return promise.future();
+  }
+  
+  public Future<IPersistentMap> send(HttpRequest<Buffer> request, Object body) {
+    Buffer buffer;
+    try {
+      buffer = toBuffer(body);
+    } catch (Throwable ex) {
+      Promise<IPersistentMap> promise = Promise.promise();
+      promise.fail(ex);
+      return promise.future();
+    }
+    return send(request, buffer);
+  }
+  
+  public Future<IPersistentMap> send(HttpRequest<Buffer> request, Buffer body) {
+    Promise<IPersistentMap> promise = Promise.promise();
+    request.sendBuffer(body, new RingResponseAdapter(promise));
+    return promise.future();
+  }
+  
+  public Future<IPersistentMap> sendForm(HttpRequest<Buffer> request, IPersistentMap body) {
+    return sendForm(request, toMultiMap(body));
+  }
+  
+  public Future<IPersistentMap> sendForm(HttpRequest<Buffer> request, MultiMap body) {
+    Promise<IPersistentMap> promise = Promise.promise();
+    request.sendForm(body, new RingResponseAdapter(promise));
+    return promise.future();
+  }
+  
+  public Future<IPersistentMap> sendMultiPartForm(HttpRequest<Buffer> request, IPersistentMap body) {
+    return sendMultiPartForm(request, toMultipartForm(body));
+  }
+  
+  public Future<IPersistentMap> sendMultiPartForm(HttpRequest<Buffer> request, MultipartForm body) {
+    Promise<IPersistentMap> promise = Promise.promise();
+    request.sendMultipartForm(body, new RingResponseAdapter(promise));
+    return promise.future();
+  }
+  
+  public void shutdown() {
+    client.close();
+  }
+  
 }
